TASK-54025 Rework Login Process to make it more extensible using Regular Vue & Vuetify Application
