+++ web/portal/src/main/resources/locale/portal/webui_ar.properties
@@ -229,27 +229,6 @@ UIGroupMembershipForm.msg.Invalid-char=\u0627\u0644\u0645\u064A\u062F\u0627\u064
 UIGroupMembershipForm.msg.duplicate-user=\u064A\u062A\u0645 \u062A\u0643\u0631\u0627\u0631 \u0628\u0639\u0636 \u0627\u0644\u0645\u0633\u062A\u062E\u062F\u0645\u064A\u0646. \u064A\u0631\u062C\u0649 \u0625\u0639\u0627\u062F\u0629 \u0627\u0644\u062A\u062B\u0628\u062A \u0646\u0645 \u0647\u0630\u0627 \u0627\u0644\u0623\u0645\u0631.
 UIGroupMembershipForm.msg.group-not-select=\u0627\u0644\u0631\u062C\u0627\u0621 \u0627\u062E\u062A\u064A\u0627\u0631 \u0645\u062C\u0645\u0648\u0639\u0629.
 UIGroupMembershipForm.msg.membership-exist=\u0627\u0644\u0645\u0633\u062A\u062E\u062F\u0645 "{0}" \u0628\u0627\u0644\u0641\u0639\u0644 \u0646\u0641\u0633 \u0623\u0639\u0636\u0627\u0621 \u0641\u064A \u0645\u062C\u0645\u0648\u0639\u0629 "{1}" \u060C         \u0641\u064A\u0631\u062C\u0649 \u062A\u062D\u062F\u064A\u062F \u0639\u0636\u0648\u064A\u0629 \u0623\u062E\u0631\u0649.       
-UILoginForm.msg.Invalid-account=\u0627\u0633\u0645 \u0627\u0644\u0645\u0633\u062A\u062E\u062F\u0645 \u0648 / \u0623\u0648 \u0643\u0644\u0645\u0629 \u0627\u0644\u0645\u0631\u0648\u0631 \u062E\u0627\u0637\u0626\u0629 \u0623\u0648 \u0641\u0627\u0631\u063A\u0629.         \u064A\u0631\u062C\u0649 \u0627\u0644\u0645\u062D\u0627\u0648\u0644\u0629 \u0645\u0631\u0629 \u0623\u062E\u0631\u0649.       
-UILoginForm.label.welcome=\u0623\u0647\u0644\u0627 \u0648\u0633\u0647\u0644\u0627 \u0628\u0643\u0645
-UILoginForm.label.login=\u062F\u062E\u0648\u0644
-UILoginForm.label.Discard=\u0646\u0628\u0630
-UILoginForm.label.user=\u0627\u0644\u0645\u0633\u062A\u062E\u062F\u0645
-UILoginForm.label.forgot=\u0646\u0633\u064A\u062A \u0627\u0633\u0645 \u0627\u0644\u0645\u0633\u062A\u062E\u062F\u0645 / \u0643\u0644\u0645\u0629 \u0627\u0644\u0633\u0631\u061F
-UILoginForm.label.password=\u0643\u0644\u0645\u0629 \u0627\u0644\u0633\u0631
-UILoginForm.label.UserName=\u0627\u0633\u0645 \u0627\u0644\u0645\u0633\u062A\u062E\u062F\u0645
-UILoginForm.label.RememberOnComputer=\u0627\u0633\u062A\u0645\u0631\u0627\u0631\u064A\u0629 \u062A\u0633\u062C\u064A\u0644 \u0627\u0644\u062F\u062E\u0648\u0644
-UILoginForm.label.NotMember=\u0644\u0633\u062A \u0639\u0636\u0648\u0627\u061F
-UILoginForm.label.Signup=\u0627\u0634\u062A\u0631\u0643
-UILoginForm.label.Signin=\u062A\u0633\u062C\u064A\u0644 \u0627\u0644\u062F\u062E\u0648\u0644
-UILoginForm.label.Signout=\u062A\u0633\u062C\u064A\u0644 \u0627\u0644\u062E\u0631\u0648\u062C
-UILoginForm.label.ForAccount=\u0633\u062C\u0644 \u0627\u0644\u0622\u0646 \u0644\u0644\u062D\u0635\u0648\u0644 \u0639\u0644\u0649 \u062D\u0633\u0627\u0628
-UILoginForm.label.SigninFail=\u0646\u062D\u0646 \u0622\u0633\u0641\u0648\u0646. \u0644\u0627 \u064A\u0645\u0643\u0646\u0643 \u062A\u0633\u062C\u064A\u0644 \u0627\u0644\u062F\u062E\u0648\u0644 \u0625\u0644\u0649 \u0627\u0644\u0645\u0646\u0635\u0629. \u064A\u0631\u062C\u0649 \u0627\u0644\u0645\u062D\u0627\u0648\u0644\u0629 \u0645\u0631\u0629 \u0623\u062E\u0631\u0649 \u0623\u0648 \u0627\u0644\u0627\u062A\u0635\u0627\u0644 \u0628\u0627\u0644\u0645\u0633\u0624\u0648\u0644.
-UILoginForm.label.DisabledUserSignin=\u0646\u062D\u0646 \u0622\u0633\u0641\u0648\u0646. \u0644\u0627 \u064A\u0645\u0643\u0646\u0643 \u062A\u0633\u062C\u064A\u0644 \u0627\u0644\u062F\u062E\u0648\u0644 \u0625\u0644\u0649 \u0627\u0644\u0645\u0646\u0635\u0629. \u064A\u0631\u062C\u0649 \u0627\u0644\u0645\u062D\u0627\u0648\u0644\u0629 \u0645\u0631\u0629 \u0623\u062E\u0631\u0649 \u0623\u0648 \u0627\u0644\u0627\u062A\u0635\u0627\u0644 \u0628\u0627\u0644\u0645\u0633\u0624\u0648\u0644.
-UILoginForm.label.SSODisabledUserSignin=\u0644\u0642\u062F \u062D\u0627\u0648\u0644\u062A \u0627\u0644\u0627\u062A\u0635\u0627\u0644 \u0628\u062D\u0633\u0627\u0628 \u0645\u0633\u062A\u062E\u062F\u0645 \u0648\u0642\u0639 \u062A\u0639\u0644\u064A\u0642\u0647.
-UILoginForm.label.SSODisabledUserSignoutRequest=\u062B\u0645 \u0642\u0645 \u0628\u062A\u0633\u062C\u064A\u0644 \u0627\u0644\u062F\u062E\u0648\u0644 \u0628\u0648\u0627\u0633\u0637\u0629 \u062D\u0633\u0627\u0628 \u0622\u062E\u0631.
-UILoginForm.label.Copyright=\u062D\u0642\u0648\u0642 \u0627\u0644\u0637\u0628\u0639 \u0648\u0627\u0644\u0646\u0634\u0631 \u00A9 2009-2016. \u062C\u0645\u064A\u0639 \u0627\u0644\u062D\u0642\u0648\u0642 \u0645\u062D\u0641\u0648\u0638\u0629\u060C \u0623\u0643\u0632\u0648 \u0628\u0644\u0627\u062A\u0641\u0648\u0631\u0645 SAS
-UILoginForm.label.SignInWith=\u062A\u0633\u062C\u064A\u0644 \u0627\u0644\u062F\u062E\u0648\u0644 \u0628\u0648\u0627\u0633\u0637\u0629
-UILoginForm.label.Delimiter=\u0623\u0648 \u062A\u0633\u062C\u064A\u0644 \u0627\u0644\u062F\u062E\u0648\u0644 \u0628\u0627\u0633\u062A\u062E\u062F\u0627\u0645
 # Password setup related
 UISetupForm.label.setup.SetPasswordTitle=\u0625\u0639\u062F\u0627\u062F \u0643\u0644\u0645\u0629 \u0645\u0631\u0648\u0631\u0627\u0644\u062E\u0627\u0635\u0629 \u0628\u0627\u0644\u0645\u0633\u062A\u062E\u062F\u0645 \u0627\u0644\u0631\u0626\u064A\u0633\u064A
 UISetupForm.label.setup.Close=\u0627\u063A\u0644\u0627\u0642
@@ -261,14 +240,6 @@ UISetupForm.label.mobile.copyright.Intro=\u062D\u0642\u0648\u0642 \u0627\u0644\u
 UISetupForm.label.mobile.copyright.RH=\u0634\u0631\u0643\u0629 \u0631\u064A\u062F\u0647\u0627\u062A
 UISetupForm.label.mobile.copyright.And=\u0648
 UISetupForm.label.mobile.copyright.Exo=\u0623\u0643\u0632\u0648 \u0628\u0644\u0627\u062A\u0641\u0648\u0631\u0645 SAS
-# Mobile login.jsp related
-UILoginForm.label.mobile.Login=\u062A\u0633\u062C\u064A\u0644 \u0627\u0644\u062F\u062E\u0648\u0644
-UILoginForm.label.mobile.copyright.Intro=\u062D\u0642\u0648\u0642 \u0627\u0644\u0637\u0628\u0639 \u0648\u0627\u0644\u0646\u0634\u0631 &copy;  2013. \u062C\u0645\u064A\u0639 \u0627\u0644\u062D\u0642\u0648\u0642 \u0645\u062D\u0641\u0648\u0638\u0629\u060C
-UILoginForm.label.mobile.copyright.RH=\u0634\u0631\u0643\u0629 \u0631\u064A\u062F\u0647\u0627\u062A
-UILoginForm.label.mobile.copyright.And=\u0648
-UILoginForm.label.mobile.copyright.Exo=\u0623\u0643\u0632\u0648 \u0628\u0644\u0627\u062A\u0641\u0648\u0631\u0645 SAS
-UILoginForm.label.mobile.login.oauth.Delimiter=\u0623\u0648 \u062A\u0633\u062C\u064A\u0644 \u0627\u0644\u062F\u062E\u0648\u0644 \u0628\u0627\u0633\u062A\u062E\u062F\u0627\u0645
-UILoginForm.label.mobile.login.Close=\u0625\u063A\u0644\u0627\u0642
 #############################################################################
 #                              Message Error                                #
 #############################################################################
