+++ web/portal/src/main/resources/locale/portal/webui_az.properties
@@ -229,27 +229,6 @@ UIGroupMembershipForm.msg.Invalid-char=The field "{0}" can only contain alpha, d
 UIGroupMembershipForm.msg.duplicate-user=Some user names are duplicated. Please verify.
 UIGroupMembershipForm.msg.group-not-select=Please select a group.
 UIGroupMembershipForm.msg.membership-exist=User "{0}" is already has  already the same role in the group "{1}", please select another role.
-UILoginForm.msg.Invalid-account=User Name and/or Password are wrong or empty. Please try again.
-UILoginForm.label.welcome=Welcome
-UILoginForm.label.login=Giri\u015F
-UILoginForm.label.Discard=Discard
-UILoginForm.label.user=\u0130stifad\u0259\u00E7i
-UILoginForm.label.forgot=Forgot your User Name/Password?
-UILoginForm.label.password=Password
-UILoginForm.label.UserName=User name
-UILoginForm.label.RememberOnComputer=Stay signed in
-UILoginForm.label.NotMember=Not a member?
-UILoginForm.label.Signup=Signup
-UILoginForm.label.Signin=Sign in
-UILoginForm.label.Signout=Sign out
-UILoginForm.label.ForAccount=Register now for an account
-UILoginForm.label.SigninFail=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.DisabledUserSignin=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.SSODisabledUserSignin=You attempted to connect to a suspended user account.
-UILoginForm.label.SSODisabledUserSignoutRequest=and then sign in with another account.
-UILoginForm.label.Copyright=Copyright &copy; 2009-2016. All rights reserved, eXo Platform SAS
-UILoginForm.label.SignInWith=Sign in with
-UILoginForm.label.Delimiter=or sign in with
 # Password setup related
 UISetupForm.label.setup.SetPasswordTitle=Root password setup
 UISetupForm.label.setup.Close=Close
@@ -261,14 +240,6 @@ UISetupForm.label.mobile.copyright.Intro=Copyright &copy; 2013. All rights reser
 UISetupForm.label.mobile.copyright.RH=Red Hat, Inc
 UISetupForm.label.mobile.copyright.And=and
 UISetupForm.label.mobile.copyright.Exo=eXo Platform SAS
-# Mobile login.jsp related
-UILoginForm.label.mobile.Login=Giri\u015F
-UILoginForm.label.mobile.copyright.Intro=Copyright &copy; 2013. All rights reserved,
-UILoginForm.label.mobile.copyright.RH=Red Hat, Inc
-UILoginForm.label.mobile.copyright.And=and
-UILoginForm.label.mobile.copyright.Exo=eXo Platform SAS
-UILoginForm.label.mobile.login.oauth.Delimiter=or sign in with
-UILoginForm.label.mobile.login.Close=close
 #############################################################################
 #                              Message Error                                #
 #############################################################################
@@ -847,17 +818,6 @@ UILogged.label.RollBack=Roll Back
 UILogged.label.Save=#{word.save}
 UILogged.label.Finish=#{word.finish}
 UILogged.label.Abort=#{word.abort}
-############################################################################
-#   org.exoplatform.portal.component.widget.UILoginForm              #
-############################################################################
-UIPortalComponentLogin.label.password=Password
-UIPortalComponentLogin.label.username=User name
-UIForgetPasswordWizard.title=Why are you are not able to login ?
-UIForgetPasswordWizard.action.Next=Next
-UIForgetPasswordWizard.action.Back=#{word.back}
-UIForgetPasswordWizard.info=We apologize for any inconvenience you're experiencing by not being able to access this website.<br />To resolve this issue as quickly as possible, please follow the troubleshooting steps below.<br /><br />1. Recover your password: enter <strong>your username</strong> then click send.<br/>2. Recover your username: enter <strong>your email address</strong> then click send.<br/>
-UIForgetPasswordWizard.label.forgotpassword=Forgot my password
-UIForgetPasswordWizard.label.forgotusername=Forgot my username
 ############################################################################
 #   org.exoplatform.webui.portal.UIForgetPassword & UIResetPassword  #
 ############################################################################
