+++ web/portal/src/main/resources/locale/portal/webui_en.properties
@@ -229,27 +229,6 @@ UIGroupMembershipForm.msg.Invalid-char=The field "{0}" can only contain alpha, d
 UIGroupMembershipForm.msg.duplicate-user=Some user names are duplicated. Please verify.
 UIGroupMembershipForm.msg.group-not-select=Please select a group.
 UIGroupMembershipForm.msg.membership-exist=User "{0}" is already has  already the same role in the group "{1}", please select another role.
-UILoginForm.msg.Invalid-account=User Name and/or Password are wrong or empty. Please try again.
-UILoginForm.label.welcome=Welcome
-UILoginForm.label.login=Login
-UILoginForm.label.Discard=Discard
-UILoginForm.label.user=User
-UILoginForm.label.forgot=Forgot your User Name/Password?
-UILoginForm.label.password=Password
-UILoginForm.label.UserName=User name
-UILoginForm.label.RememberOnComputer=Stay signed in
-UILoginForm.label.NotMember=Not a member?
-UILoginForm.label.Signup=Signup
-UILoginForm.label.Signin=Sign in
-UILoginForm.label.Signout=Sign out
-UILoginForm.label.ForAccount=Register now for an account
-UILoginForm.label.SigninFail=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.DisabledUserSignin=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.SSODisabledUserSignin=You attempted to connect to a suspended user account.
-UILoginForm.label.SSODisabledUserSignoutRequest=and then sign in with another account.
-UILoginForm.label.Copyright=Copyright &copy; 2009-2016. All rights reserved, eXo Platform SAS
-UILoginForm.label.SignInWith=Sign in with
-UILoginForm.label.Delimiter=or sign in with
 # Password setup related
 UISetupForm.label.setup.SetPasswordTitle=Root password setup
 UISetupForm.label.setup.Close=Close
