+++ web/portal/src/main/resources/locale/portal/webui_es_ES.properties
@@ -229,27 +229,6 @@ UIGroupMembershipForm.msg.Invalid-char=El campo "{0}" s\u00F3lo admite caractere
 UIGroupMembershipForm.msg.duplicate-user=Alg\u00FAn nombre de usuario est\u00E1 duplicado. Por favor, verif\u00EDquelo.
 UIGroupMembershipForm.msg.group-not-select=Seleccione un Grupo.
 UIGroupMembershipForm.msg.membership-exist=El usuario {0} ya es tiene ese rol en el grupo {1}. Selecciona otro rol.
-UILoginForm.msg.Invalid-account=El nombre de Usuario y/o la Cantrase\u00F1a son incorrectas o est\u00E1n vac\u00EDas. Por favor, pruebe de nuevo.
-UILoginForm.label.welcome=Bienvenido
-UILoginForm.label.login=Iniciar sesi\u00F3n
-UILoginForm.label.Discard=Descartar
-UILoginForm.label.user=Usuario
-UILoginForm.label.forgot=\u00BFOlvidaste tu Nombre de Usuario/Contrase\u00F1a?
-UILoginForm.label.password=Contrase\u00F1a
-UILoginForm.label.UserName=Nombre de usuario
-UILoginForm.label.RememberOnComputer=Mantener la sesi\u00F3n iniciada
-UILoginForm.label.NotMember=\u00BFNo es Miembro?
-UILoginForm.label.Signup=Reg\u00EDstrate
-UILoginForm.label.Signin=Inicio de sesi\u00F3n
-UILoginForm.label.Signout=desconectar
-UILoginForm.label.ForAccount=Registarse ahora para crear una cuenta
-UILoginForm.label.SigninFail=Lo sentimos. Error al iniciar sesi\u00F3n en la plataforma. Reint\u00E9ntalo o contacta con tu administrador.
-UILoginForm.label.DisabledUserSignin=Lo sentimos. Error al iniciar sesi\u00F3n en la plataforma. Reint\u00E9ntalo o contacta con tu administrador.
-UILoginForm.label.SSODisabledUserSignin=Ha intentado conectarse a una cuenta de usuario suspendido.
-UILoginForm.label.SSODisabledUserSignoutRequest=y luego inicia sesi\u00F3n con otra cuenta.
-UILoginForm.label.Copyright=Copyright &copy; 2009-2016. Todos los derechos reservados, eXo Platform SAS
-UILoginForm.label.SignInWith=Ingresar con
-UILoginForm.label.Delimiter=o iniciar sesi\u00F3n con
 # Password setup related
 UISetupForm.label.setup.SetPasswordTitle=Configuraci\u00F3n de la contrase\u00F1a de root
 UISetupForm.label.setup.Close=Cerrar
@@ -261,14 +240,6 @@ UISetupForm.label.mobile.copyright.Intro=Copyright &copy; 2013. Todos los derech
 UISetupForm.label.mobile.copyright.RH=Red Hat, Inc.
 UISetupForm.label.mobile.copyright.And=y
 UISetupForm.label.mobile.copyright.Exo=eXo Platform SAS
-# Mobile login.jsp related
-UILoginForm.label.mobile.Login=Inicio de sesi\u00F3n
-UILoginForm.label.mobile.copyright.Intro=Copyright &copy; 2013. Todos los derechos reservados,
-UILoginForm.label.mobile.copyright.RH=Red Hat, Inc.
-UILoginForm.label.mobile.copyright.And=y
-UILoginForm.label.mobile.copyright.Exo=eXo Platform SAS
-UILoginForm.label.mobile.login.oauth.Delimiter=o iniciar sesi\u00F3n con
-UILoginForm.label.mobile.login.Close=cerrar
 #############################################################################
 #                              Message Error                                #
 #############################################################################
@@ -847,17 +818,6 @@ UILogged.label.RollBack=Deshacer
 UILogged.label.Save=#{word.save}
 UILogged.label.Finish=#{word.finish}
 UILogged.label.Abort=#{word.abort}
-############################################################################
-#   org.exoplatform.portal.component.widget.UILoginForm              #
-############################################################################
-UIPortalComponentLogin.label.password=Contrase\u00F1a
-UIPortalComponentLogin.label.username=Nombre de Usuario
-UIForgetPasswordWizard.title=\u00BFPorqu\u00E9 no puede conectarse ?
-UIForgetPasswordWizard.action.Next=Siguiente
-UIForgetPasswordWizard.action.Back=#{word.back}
-UIForgetPasswordWizard.info=Sentimos cualquier inconveniente por no poder acceder a eXo.<br />Para resolver esta incidencia tan pronto como sea posible, por favor sigue los pasos de resoluci\u00F3n de incidencias que siguen.<br /><br />1. Recuperar tu Contrase\u00F1a: introduce <strong>tu nombre de usuario</strong> y pulsa enviar.<br/>2. Recuperar tu nombre de usuario: introduce <strong>tu correo electr\u00F3nico</strong> y pulsa enviar.<br/>
-UIForgetPasswordWizard.label.forgotpassword=Olvid\u00E9 mi Contrase\u00F1a
-UIForgetPasswordWizard.label.forgotusername=Olvid\u00E9 mi Nombre de Usuario
 ############################################################################
 #   org.exoplatform.webui.portal.UIForgetPassword & UIResetPassword  #
 ############################################################################
