+++ web/portal/src/main/resources/locale/portal/webui_fr.properties
@@ -229,27 +229,6 @@ UIGroupMembershipForm.msg.Invalid-char=Le champ "{0}" ne doit contenir que des l
 UIGroupMembershipForm.msg.duplicate-user=Certains noms d'utilisateurs sont en double. V\u00E9rifiez votre saisie.
 UIGroupMembershipForm.msg.group-not-select=Veuillez s\u00E9lectionner un groupe.
 UIGroupMembershipForm.msg.membership-exist=L'utilisateur "{0}" a d\u00E9j\u00E0 le m\u00EAme r\u00F4le dans le groupe "{1}, veuillez s\u00E9lectionner un autre r\u00F4le.
-UILoginForm.msg.Invalid-account=Nom d'utilisateur et/ou mot de passe incorrect ou manquant. Veuillez r\u00E9-essayer.
-UILoginForm.label.welcome=Bienvenue
-UILoginForm.label.login=Se connecter
-UILoginForm.label.Discard=Annuler
-UILoginForm.label.user=Utilisateur
-UILoginForm.label.forgot=Identifiant / mot de passe oubli\u00E9 ?
-UILoginForm.label.password=Mot de passe
-UILoginForm.label.UserName=Identifiant
-UILoginForm.label.RememberOnComputer=Rester connect\u00E9
-UILoginForm.label.NotMember=Pas encore membre?
-UILoginForm.label.Signup=Enregistrement
-UILoginForm.label.Signin=Connexion
-UILoginForm.label.Signout=D\u00E9connexion
-UILoginForm.label.ForAccount=Cr\u00E9er un compte
-UILoginForm.label.SigninFail=Nous sommes d\u00E9sol\u00E9s. Vous ne pouvez pas vous connecter \u00E0 la plateforme. Veuillez essayer une nouvelle fois ou bien contacter votre administrateur.
-UILoginForm.label.DisabledUserSignin=Nous sommes d\u00E9sol\u00E9s. Vous ne pouvez pas vous connecter \u00E0 la plateforme. Veuillez essayer une nouvelle fois ou bien contacter votre administrateur.
-UILoginForm.label.SSODisabledUserSignin=Vous avez essay\u00E9 de vous connecter \u00E0 un compte utilisateur suspendu.
-UILoginForm.label.SSODisabledUserSignoutRequest=et puis connectez-vous avec un autre compte.
-UILoginForm.label.Copyright=Copyright &copy; 2009-2016. Tous droits r\u00E9serv\u00E9s, eXo Platform SAS
-UILoginForm.label.SignInWith=Connexion avec
-UILoginForm.label.Delimiter=ou connectez-vous avec
 # Password setup related
 UISetupForm.label.setup.SetPasswordTitle=Configuration du mot de passe root
 UISetupForm.label.setup.Close=Fermer
@@ -261,14 +240,6 @@ UISetupForm.label.mobile.copyright.Intro=Copyright &copy; 2013. All rights reser
 UISetupForm.label.mobile.copyright.RH=Red Hat, Inc
 UISetupForm.label.mobile.copyright.And=et
 UISetupForm.label.mobile.copyright.Exo=eXo Platform SAS
-# Mobile login.jsp related
-UILoginForm.label.mobile.Login=Connexion
-UILoginForm.label.mobile.copyright.Intro=Copyright &copy; 2013. Tous droits r\u00E9serv\u00E9s,
-UILoginForm.label.mobile.copyright.RH=Red Hat, Inc
-UILoginForm.label.mobile.copyright.And=et
-UILoginForm.label.mobile.copyright.Exo=eXo Platform SAS
-UILoginForm.label.mobile.login.oauth.Delimiter=ou connectez-vous avec
-UILoginForm.label.mobile.login.Close=Fermer
 #############################################################################
 #                              Message Error                                #
 #############################################################################
@@ -847,17 +818,6 @@ UILogged.label.RollBack=Annuler les modifications en cours
 UILogged.label.Save=#{word.save}
 UILogged.label.Finish=#{word.finish}
 UILogged.label.Abort=#{word.abort}
-############################################################################
-#   org.exoplatform.portal.component.widget.UILoginForm              #
-############################################################################
-UIPortalComponentLogin.label.password=Mot de passe
-UIPortalComponentLogin.label.username=Identifiant
-UIForgetPasswordWizard.title=Pourquoi ne pouvez-vous pas vous connecter ?
-UIForgetPasswordWizard.action.Next=Suivant
-UIForgetPasswordWizard.action.Back=#{word.back}
-UIForgetPasswordWizard.info=Nous sommes d\u00E9sol\u00E9 que vous ne puissiez acc\u00E9der \u00E0 ce site.<br />Afin de rem\u00E9dier \u00E0 ce probl\u00E8me, merci de suivre une des deux actions suivantes.<br />En suivant ces instructions vous pourrez vous connecter de nouveau.<br /><br />1. R\u00E9cup\u00E9rer votre mot de passe \u00E0 partir de <strong>votre identifiant</strong>.<br/>2. R\u00E9cup\u00E9rer votre identifiant \u00E0 partir de <strong>votre adresse email</strong><br/>
-UIForgetPasswordWizard.label.forgotpassword=Oubli de mon mot de passe
-UIForgetPasswordWizard.label.forgotusername=Oubli de mon identifiant
 ############################################################################
 #   org.exoplatform.webui.portal.UIForgetPassword & UIResetPassword  #
 ############################################################################
