+++ web/portal/src/main/resources/locale/portal/webui_lt.properties
@@ -229,27 +229,6 @@ UIGroupMembershipForm.msg.Invalid-char=Lauke "{0}" gali b\u016Bti tik raid\u0117
 UIGroupMembershipForm.msg.duplicate-user=Some user names are duplicated. Please verify.
 UIGroupMembershipForm.msg.group-not-select=Pasirinkite grup\u0119.
 UIGroupMembershipForm.msg.membership-exist=User "{0}" is already has  already the same role in the group "{1}", please select another role.
-UILoginForm.msg.Invalid-account=Vartotojo vardas ir/ar slapta\u017Eodis yra neteisingas arba tu\u0161\u010Dias. Bandykite dar kart\u0105.
-UILoginForm.label.welcome=Sveiki
-UILoginForm.label.login=Prisijungti
-UILoginForm.label.Discard=Atmesti
-UILoginForm.label.user=Vartotojas
-UILoginForm.label.forgot=Pamir\u0161ote savo vartotojo vard\u0105 ir slapta\u017Eod\u012F?
-UILoginForm.label.password=Slapta\u017Eodis
-UILoginForm.label.UserName=Vartotojo vardas
-UILoginForm.label.RememberOnComputer=Stay signed in
-UILoginForm.label.NotMember=Ne narys?
-UILoginForm.label.Signup=Registruotis
-UILoginForm.label.Signin=Prisijungti
-UILoginForm.label.Signout=Atsijungti
-UILoginForm.label.ForAccount=U\u017Esiregistruoti paskyr\u0105 dabar
-UILoginForm.label.SigninFail=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.DisabledUserSignin=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.SSODisabledUserSignin=You attempted to connect to a suspended user account.
-UILoginForm.label.SSODisabledUserSignoutRequest=and then sign in with another account.
-UILoginForm.label.Copyright=Copyright &copy; 2009-2016. All rights reserved, eXo Platform SAS
-UILoginForm.label.SignInWith=Sign in with
-UILoginForm.label.Delimiter=or sign in with
 # Password setup related
 UISetupForm.label.setup.SetPasswordTitle=Root password setup
 UISetupForm.label.setup.Close=U\u017Edaryti
@@ -261,14 +240,6 @@ UISetupForm.label.mobile.copyright.Intro=Copyright &copy; 2013. All rights reser
 UISetupForm.label.mobile.copyright.RH=Red Hat, Inc
 UISetupForm.label.mobile.copyright.And=and
 UISetupForm.label.mobile.copyright.Exo=eXo Platform SAS
-# Mobile login.jsp related
-UILoginForm.label.mobile.Login=Prisijungti
-UILoginForm.label.mobile.copyright.Intro=Copyright &copy; 2013. All rights reserved,
-UILoginForm.label.mobile.copyright.RH=Red Hat, Inc
-UILoginForm.label.mobile.copyright.And=and
-UILoginForm.label.mobile.copyright.Exo=eXo Platform SAS
-UILoginForm.label.mobile.login.oauth.Delimiter=or sign in with
-UILoginForm.label.mobile.login.Close=close
 #############################################################################
 #                              Message Error                                #
 #############################################################################
@@ -847,17 +818,6 @@ UILogged.label.RollBack=Gra\u017Einti atgal
 UILogged.label.Save=#{word.save}
 UILogged.label.Finish=#{word.finish}
 UILogged.label.Abort=#{word.abort}
-############################################################################
-#   org.exoplatform.portal.component.widget.UILoginForm              #
-############################################################################
-UIPortalComponentLogin.label.password=Slapta\u017Eodis
-UIPortalComponentLogin.label.username=Vartotojo vardas
-UIForgetPasswordWizard.title=Kod\u0117l tu negali prisijungti?
-UIForgetPasswordWizard.action.Next=Kitas
-UIForgetPasswordWizard.action.Back=#{word.back}
-UIForgetPasswordWizard.info=Atsipra\u0161ome u\u017E nepatogumus, kuriuos patiriate negal\u0117dami pasiekti \u0161ios svetain\u0117s. <br /> Kad kuo grei\u010Diau i\u0161spr\u0119sti \u0161i\u0105 problem\u0105, atlikite toliau pateiktus trik\u010Di\u0173 \u0161alinimo veiksmus. <br /> <br />1. Atkurti slapta\u017Eod\u012F: \u012Fveskite <strong>savo vartotojo vard\u0105</strong>, tada spustel\u0117kite Si\u0173sti.<br/>2. Atkurti j\u016Bs\u0173 vartotojo vard\u0105: \u012Fveskite <strong>savo elektroninio pa\u0161to adres\u0105</strong> ir spustel\u0117kite si\u0173sti. <br/>
-UIForgetPasswordWizard.label.forgotpassword=Pamir\u0161au slapta\u017Eod\u012F
-UIForgetPasswordWizard.label.forgotusername=Pamir\u0161au savo vartotojo vard\u0105
 ############################################################################
 #   org.exoplatform.webui.portal.UIForgetPassword & UIResetPassword  #
 ############################################################################
