+++ web/portal/src/main/resources/locale/portal/webui_no.properties
@@ -229,27 +229,6 @@ UIGroupMembershipForm.msg.Invalid-char=Feltet "{0}" kan bare inneholde bokstaver
 UIGroupMembershipForm.msg.duplicate-user=Some user names are duplicated. Please verify.
 UIGroupMembershipForm.msg.group-not-select=Velg en gruppe.
 UIGroupMembershipForm.msg.membership-exist=User "{0}" is already has  already the same role in the group "{1}", please select another role.
-UILoginForm.msg.Invalid-account=Brukernavn og/eller passordet er feil. Pr\u00F8v igjen.
-UILoginForm.label.welcome=Velkommen
-UILoginForm.label.login=Innlogging
-UILoginForm.label.Discard=Forkast
-UILoginForm.label.user=Bruker
-UILoginForm.label.forgot=Har du glemt brukernavn/passord?
-UILoginForm.label.password=Passord
-UILoginForm.label.UserName=Brukernavn
-UILoginForm.label.RememberOnComputer=Stay signed in
-UILoginForm.label.NotMember=Ikke medlem?
-UILoginForm.label.Signup=Registrer deg
-UILoginForm.label.Signin=Logg inn
-UILoginForm.label.Signout=Logg ut
-UILoginForm.label.ForAccount=Registrer deg n\u00E5 for en konto
-UILoginForm.label.SigninFail=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.DisabledUserSignin=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.SSODisabledUserSignin=You attempted to connect to a suspended user account.
-UILoginForm.label.SSODisabledUserSignoutRequest=and then sign in with another account.
-UILoginForm.label.Copyright=Copyright &copy; 2009-2016. All rights reserved, eXo Platform SAS
-UILoginForm.label.SignInWith=Sign in with
-UILoginForm.label.Delimiter=or sign in with
 # Password setup related
 UISetupForm.label.setup.SetPasswordTitle=Root password setup
 UISetupForm.label.setup.Close=Lukk
@@ -261,14 +240,6 @@ UISetupForm.label.mobile.copyright.Intro=Copyright &copy; 2013. All rights reser
 UISetupForm.label.mobile.copyright.RH=Red Hat, Inc
 UISetupForm.label.mobile.copyright.And=and
 UISetupForm.label.mobile.copyright.Exo=eXo Platform SAS
-# Mobile login.jsp related
-UILoginForm.label.mobile.Login=Innlogging
-UILoginForm.label.mobile.copyright.Intro=Copyright &copy; 2013. All rights reserved,
-UILoginForm.label.mobile.copyright.RH=Red Hat, Inc
-UILoginForm.label.mobile.copyright.And=and
-UILoginForm.label.mobile.copyright.Exo=eXo Platform SAS
-UILoginForm.label.mobile.login.oauth.Delimiter=or sign in with
-UILoginForm.label.mobile.login.Close=close
 #############################################################################
 #                              Message Error                                #
 #############################################################################
@@ -847,17 +818,6 @@ UILogged.label.RollBack=Tilbakerulling
 UILogged.label.Save=#{word.save}
 UILogged.label.Finish=#{word.finish}
 UILogged.label.Abort=#{word.abort}
-############################################################################
-#   org.exoplatform.portal.component.widget.UILoginForm              #
-############################################################################
-UIPortalComponentLogin.label.password=Passord
-UIPortalComponentLogin.label.username=Brukernavn
-UIForgetPasswordWizard.title=Hvorfor kan du ikke logge inn?
-UIForgetPasswordWizard.action.Next=Neste
-UIForgetPasswordWizard.action.Back=#{word.back}
-UIForgetPasswordWizard.info=Vi beklager alt bry du opplever ved ikke \u00E5 f\u00E5 tilgang til dette nettstedet. <br/> For \u00E5 l\u00F8se problemet s\u00E5 raskt som mulig, kan du f\u00F8lge feils\u00F8kingstrinnene nedenfor. <br/> <br/> 1. Gjenopprette passord: Skriv inn <strong>ditt brukernavn</strong> og klikk send. <br/>2. Gjenopprette brukernavn: oppgi <strong>e-postadressen din</strong> og klikk send. <br/>
-UIForgetPasswordWizard.label.forgotpassword=Glemt passordet mitt
-UIForgetPasswordWizard.label.forgotusername=Glemt brukernavn
 ############################################################################
 #   org.exoplatform.webui.portal.UIForgetPassword & UIResetPassword  #
 ############################################################################
