+++ web/portal/src/main/resources/locale/portal/webui_sl.properties
@@ -229,27 +229,6 @@ UIGroupMembershipForm.msg.Invalid-char=Polje "{0}" lahko vsebuje samo alfanueri\
 UIGroupMembershipForm.msg.duplicate-user=Nekateri uporabniki so podvojeni. Prosim, preverite.
 UIGroupMembershipForm.msg.group-not-select=Izberite skupino.
 UIGroupMembershipForm.msg.membership-exist=User "{0}" is already has  already the same role in the group "{1}", please select another role.
-UILoginForm.msg.Invalid-account=Uporabni\u0161ko ime in/ali geslo je napa\u010Dno ali prazno. Prosimo, poskusite ponovno.
-UILoginForm.label.welcome=Dobrodo\u0161li
-UILoginForm.label.login=Prijava
-UILoginForm.label.Discard=Zavrzi
-UILoginForm.label.user=Uporabnik
-UILoginForm.label.forgot=Ste pozabili uporabni\u0161ko ime/geslo?
-UILoginForm.label.password=Geslo
-UILoginForm.label.UserName=Uporabni\u0161ko ime
-UILoginForm.label.RememberOnComputer=Ostani prijavljen
-UILoginForm.label.NotMember=Niste \u010Dlan?
-UILoginForm.label.Signup=Prijava
-UILoginForm.label.Signin=Vpis
-UILoginForm.label.Signout=Izpis
-UILoginForm.label.ForAccount=Registrirajte se zdaj za ra\u010Dun
-UILoginForm.label.SigninFail=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.DisabledUserSignin=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.SSODisabledUserSignin=Poskusili ste se povezati z za\u010Dasno prekinjenim uporabni\u0161kim ra\u010Dunom.
-UILoginForm.label.SSODisabledUserSignoutRequest=in se nato vpi\u0161ite z drugim ra\u010Dunom.
-UILoginForm.label.Copyright=Avtorske pravice &copy; 2009-2016. Vse pravice pridr\u017Eane, eXo platforma SAS
-UILoginForm.label.SignInWith=Vpis z
-UILoginForm.label.Delimiter=ali vpis z
 # Password setup related
 UISetupForm.label.setup.SetPasswordTitle=Nastavitev korenskega gesla
 UISetupForm.label.setup.Close=Zapri
@@ -261,14 +240,6 @@ UISetupForm.label.mobile.copyright.Intro=Copyright &copy; 2013. Vse pravice prid
 UISetupForm.label.mobile.copyright.RH=Red Hat, Inc
 UISetupForm.label.mobile.copyright.And=in
 UISetupForm.label.mobile.copyright.Exo=eXo platforma SAS
-# Mobile login.jsp related
-UILoginForm.label.mobile.Login=Prijava
-UILoginForm.label.mobile.copyright.Intro=Copyright &copy; 2013. Vse pravice pridr\u017Eane,
-UILoginForm.label.mobile.copyright.RH=Red Hat, Inc
-UILoginForm.label.mobile.copyright.And=in
-UILoginForm.label.mobile.copyright.Exo=eXo platforma SAS
-UILoginForm.label.mobile.login.oauth.Delimiter=ali vpis z
-UILoginForm.label.mobile.login.Close=zapri
 #############################################################################
 #                              Message Error                                #
 #############################################################################
@@ -847,17 +818,6 @@ UILogged.label.RollBack=Povrnitev prej\u0161njega stanja
 UILogged.label.Save=#{word.save}
 UILogged.label.Finish=#{word.finish}
 UILogged.label.Abort=#{word.abort}
-############################################################################
-#   org.exoplatform.portal.component.widget.UILoginForm              #
-############################################################################
-UIPortalComponentLogin.label.password=Geslo
-UIPortalComponentLogin.label.username=Uporabni\u0161ko ime
-UIForgetPasswordWizard.title=Zakaj se ne morete prijaviti?
-UIForgetPasswordWizard.action.Next=Naslednji
-UIForgetPasswordWizard.action.Back=#{word.back}
-UIForgetPasswordWizard.info=Opravi\u010Dujemo se vam za morebitne nev\u0161e\u010Dnosti, ker ne morete dostopati do spletne strani. <br />Da bi \u010Dimprej odpravili te\u017Eavo, sledite spodaj navedenim korakom za odpravljanje te\u017Eav. <br /> <br />1. za ponastavitev gesla: vnesite <strong>svoje uporabni\u0161ko ime</strong> in kliknite Po\u0161lji.<br/>2. za pozabljeno uporabni\u0161ko ime: vnesite <strong>va\u0161 e-po\u0161tni naslov</strong> in kliknite Po\u0161lji.<br/>
-UIForgetPasswordWizard.label.forgotpassword=Pozabljeno geslo
-UIForgetPasswordWizard.label.forgotusername=Pozabljeno uporabni\u0161ko ime
 ############################################################################
 #   org.exoplatform.webui.portal.UIForgetPassword & UIResetPassword  #
 ############################################################################
