+++ web/portal/src/main/resources/locale/portal/webui_sv_SE.properties
@@ -229,27 +229,6 @@ UIGroupMembershipForm.msg.Invalid-char=F\u00E4ltet {0} kan endast inneh\u00E5lla
 UIGroupMembershipForm.msg.duplicate-user=Some user names are duplicated. Please verify.
 UIGroupMembershipForm.msg.group-not-select=V\u00E4lj en grupp.
 UIGroupMembershipForm.msg.membership-exist=User "{0}" is already has  already the same role in the group "{1}", please select another role.
-UILoginForm.msg.Invalid-account=Anv\u00E4ndarnamnet eller l\u00F6senordet \u00E4r fel eller tomt. F\u00F6rs\u00F6k igen.
-UILoginForm.label.welcome=V\u00E4lkommen
-UILoginForm.label.login=Logga in
-UILoginForm.label.Discard=Ignorera
-UILoginForm.label.user=Anv\u00E4ndare
-UILoginForm.label.forgot=Gl\u00F6mt ditt anv\u00E4ndarnamn/l\u00F6senord?
-UILoginForm.label.password=L\u00F6senord
-UILoginForm.label.UserName=Anv\u00E4ndarnamn
-UILoginForm.label.RememberOnComputer=Stay signed in
-UILoginForm.label.NotMember=Inte medlem?
-UILoginForm.label.Signup=Registrera dig
-UILoginForm.label.Signin=Logga in
-UILoginForm.label.Signout=Logga ut
-UILoginForm.label.ForAccount=Registrera dig nu f\u00F6r ett konto
-UILoginForm.label.SigninFail=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.DisabledUserSignin=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.SSODisabledUserSignin=You attempted to connect to a suspended user account.
-UILoginForm.label.SSODisabledUserSignoutRequest=and then sign in with another account.
-UILoginForm.label.Copyright=Copyright &copy; 2009-2016. All rights reserved, eXo Platform SAS
-UILoginForm.label.SignInWith=Sign in with
-UILoginForm.label.Delimiter=or sign in with
 # Password setup related
 UISetupForm.label.setup.SetPasswordTitle=Root password setup
 UISetupForm.label.setup.Close=St\u00E4ng
@@ -261,14 +240,6 @@ UISetupForm.label.mobile.copyright.Intro=Copyright &copy; 2013. All rights reser
 UISetupForm.label.mobile.copyright.RH=Red Hat, Inc
 UISetupForm.label.mobile.copyright.And=and
 UISetupForm.label.mobile.copyright.Exo=eXo Platform SAS
-# Mobile login.jsp related
-UILoginForm.label.mobile.Login=Logga in
-UILoginForm.label.mobile.copyright.Intro=Copyright &copy; 2013. All rights reserved,
-UILoginForm.label.mobile.copyright.RH=Red Hat, Inc
-UILoginForm.label.mobile.copyright.And=and
-UILoginForm.label.mobile.copyright.Exo=eXo Platform SAS
-UILoginForm.label.mobile.login.oauth.Delimiter=or sign in with
-UILoginForm.label.mobile.login.Close=close
 #############################################################################
 #                              Message Error                                #
 #############################################################################
@@ -847,17 +818,6 @@ UILogged.label.RollBack=Rulla tillbaka
 UILogged.label.Save=#{word.save}
 UILogged.label.Finish=#{word.finish}
 UILogged.label.Abort=#{word.abort}
-############################################################################
-#   org.exoplatform.portal.component.widget.UILoginForm              #
-############################################################################
-UIPortalComponentLogin.label.password=L\u00F6senord
-UIPortalComponentLogin.label.username=Anv\u00E4ndarnamn
-UIForgetPasswordWizard.title=Varf\u00F6r kan du inte logga in?
-UIForgetPasswordWizard.action.Next=N\u00E4sta
-UIForgetPasswordWizard.action.Back=#{word.back}
-UIForgetPasswordWizard.info=Vi ber om urs\u00E4kt f\u00F6r eventuella besv\u00E4r du upplever av att inte kunna f\u00E5 tillg\u00E5ng till denna webbplats. < br / > f\u00F6r att l\u00F6sa problemet s\u00E5 snabbt som m\u00F6jligt, F\u00F6lj fels\u00F6kningsstegen nedan. < br / > < br / > 1. \u00C5terst\u00E4lla ditt l\u00F6senord: ange <strong>ditt anv\u00E4ndarnamn</strong> och klickar p\u00E5 Skicka. <br/>2. \u00C5terst\u00E4lla ditt anv\u00E4ndarnamn: ange <strong>din e-postadress</strong> och klickar p\u00E5 Skicka. <br/>
-UIForgetPasswordWizard.label.forgotpassword=Har gl\u00F6mt mitt l\u00F6senord
-UIForgetPasswordWizard.label.forgotusername=Gl\u00F6mt mitt anv\u00E4ndarnamn
 ############################################################################
 #   org.exoplatform.webui.portal.UIForgetPassword & UIResetPassword  #
 ############################################################################
