+++ web/portal/src/main/resources/locale/portal/webui_zh_TW.properties
@@ -229,27 +229,6 @@ UIGroupMembershipForm.msg.Invalid-char=\u6B04\u4F4D"{0}"\u53EA\u80FD\u5305\u542B
 UIGroupMembershipForm.msg.duplicate-user=Some user names are duplicated. Please verify.
 UIGroupMembershipForm.msg.group-not-select=\u8ACB\u9078\u64C7\u4E00\u500B\u7D44\u3002
 UIGroupMembershipForm.msg.membership-exist=User "{0}" is already has  already the same role in the group "{1}", please select another role.
-UILoginForm.msg.Invalid-account=\u4F7F\u7528\u8005\u540D\u548C/\u6216\u5BC6\u78BC\u662F\u932F\u8AA4\u7684\u6216\u70BA\u7A7A\u3002\u8ACB\u518D\u8A66\u4E00\u6B21\u3002
-UILoginForm.label.welcome=\u6B61\u8FCE
-UILoginForm.label.login=\u767B\u9304
-UILoginForm.label.Discard=\u6368\u68C4
-UILoginForm.label.user=\u4F7F\u7528\u8005
-UILoginForm.label.forgot=\u5FD8\u8A18\u4F7F\u7528\u8005\u540D/\u5BC6\u78BC\uFF1F
-UILoginForm.label.password=\u5BC6\u78BC
-UILoginForm.label.UserName=\u4F7F\u7528\u8005\u540D
-UILoginForm.label.RememberOnComputer=Stay signed in
-UILoginForm.label.NotMember=\u9084\u4E0D\u662F\u6703\u54E1\uFF1F
-UILoginForm.label.Signup=\u8A3B\u518A
-UILoginForm.label.Signin=\u767B\u9304
-UILoginForm.label.Signout=\u767B\u51FA
-UILoginForm.label.ForAccount=\u7ACB\u5373\u8A3B\u518A\u4E00\u500B\u5E33\u6236
-UILoginForm.label.SigninFail=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.DisabledUserSignin=We are sorry. You cannot log in to the platform. Please, try again or contact your administrator.
-UILoginForm.label.SSODisabledUserSignin=You attempted to connect to a suspended user account.
-UILoginForm.label.SSODisabledUserSignoutRequest=and then sign in with another account.
-UILoginForm.label.Copyright=Copyright &copy; 2009-2016. All rights reserved, eXo Platform SAS
-UILoginForm.label.SignInWith=Sign in with
-UILoginForm.label.Delimiter=or sign in with
 # Password setup related
 UISetupForm.label.setup.SetPasswordTitle=Root password setup
 UISetupForm.label.setup.Close=\u95DC\u9589
@@ -261,14 +240,6 @@ UISetupForm.label.mobile.copyright.Intro=Copyright &copy; 2013. All rights reser
 UISetupForm.label.mobile.copyright.RH=Red Hat, Inc
 UISetupForm.label.mobile.copyright.And=and
 UISetupForm.label.mobile.copyright.Exo=eXo Platform SAS
-# Mobile login.jsp related
-UILoginForm.label.mobile.Login=\u767B\u9304
-UILoginForm.label.mobile.copyright.Intro=Copyright &copy; 2013. All rights reserved,
-UILoginForm.label.mobile.copyright.RH=Red Hat, Inc
-UILoginForm.label.mobile.copyright.And=and
-UILoginForm.label.mobile.copyright.Exo=eXo Platform SAS
-UILoginForm.label.mobile.login.oauth.Delimiter=or sign in with
-UILoginForm.label.mobile.login.Close=close
 #############################################################################
 #                              Message Error                                #
 #############################################################################
@@ -847,17 +818,6 @@ UILogged.label.RollBack=\u9000\u56DE
 UILogged.label.Save=#{word.save}
 UILogged.label.Finish=#{word.finish}
 UILogged.label.Abort=#{word.abort}
-############################################################################
-#   org.exoplatform.portal.component.widget.UILoginForm              #
-############################################################################
-UIPortalComponentLogin.label.password=\u5BC6\u78BC
-UIPortalComponentLogin.label.username=\u4F7F\u7528\u8005
-UIForgetPasswordWizard.title=\u60A8\u70BA\u4EC0\u9EBC\u7121\u6CD5\u767B\u9304\uFF1F
-UIForgetPasswordWizard.action.Next=\u4E0B\u4E00\u6B65
-UIForgetPasswordWizard.action.Back=#{word.back}
-UIForgetPasswordWizard.info=\u6211\u5011\u6DF1\u8868\u6B49\u610F\u7684\u4EFB\u4F55\u4E0D\u4FBF\u8868\u793A\u4F60\u6B63\u5728\u7D93\u6B77\u7531\u4E0D\u80FD\u8A2A\u554F\u6B64\u7DB2\u7AD9\u3002 < br / > \u8981\u76E1\u53EF\u80FD\u5FEB\u5730\u89E3\u6C7A\u6B64\u554F\u984C\uFF0C\u8ACB\u6309\u7167\u4E0B\u9762\u7684\u6545\u969C\u6392\u9664\u6B65\u9A5F\u3002 < br / > < br / > 1\u3002\u6062\u5FA9\u60A8\u7684\u5BC6\u78BC\uFF1A \u8ACB\u8F38\u5165 <strong>\u60A8\u7684\u4F7F\u7528\u8005\u540D\u7A31</strong>\uFF0C\u7136\u5F8C\u6309\u4E00\u4E0B\u767C\u9001\u3002 <br/>2.\u6062\u5FA9\u60A8\u7684\u4F7F\u7528\u8005\u540D\u7A31\uFF1A \u8F38\u5165 <strong>\u60A8\u7684\u96FB\u5B50\u90F5\u4EF6\u5730\u5740</strong>\uFF0C\u7136\u5F8C\u6309\u4E00\u4E0B\u767C\u9001\u3002 <br/>
-UIForgetPasswordWizard.label.forgotpassword=\u5FD8\u8A18\u6211\u7684\u5BC6\u78BC
-UIForgetPasswordWizard.label.forgotusername=\u5FD8\u8A18\u6211\u7684\u4F7F\u7528\u8005\u540D
 ############################################################################
 #   org.exoplatform.webui.portal.UIForgetPassword & UIResetPassword  #
 ############################################################################
