--- web/portal/src/main/webapp/WEB-INF/conf/organization/picketlink-idm/infinispan-cluster.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (C) 2017 eXo Platform SAS.
-
-    This is free software; you can redistribute it and/or modify it
-    under the terms of the GNU Lesser General Public License as
-    published by the Free Software Foundation; either version 2.1 of
-    the License, or (at your option) any later version.
-
-    This software is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-    Lesser General Public License for more details.
-
-    You should have received a copy of the GNU Lesser General Public
-    License along with this software; if not, write to the Free
-    Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
-    02110-1301 USA, or see the FSF site: http://www.fsf.org.
-
--->
-<infinispan
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="urn:infinispan:config:8.0 http://www.infinispan.org/schemas/infinispan-config-8.2.xsd"
-        xmlns="urn:infinispan:config:8.0">
-
-  <jgroups>
-    <stack-file name="stack" path="jgroups/gatein-${gatein.default.jgroups.stack:udp}-ispn.xml"/>
-  </jgroups>
-
-  <cache-container  default-cache="xml-configured-cache" expiration-executor="infinispan-expiration"  listener-executor="infinispan-listener"
-                    statistics="true"  shutdown-hook="REGISTER">
-    <jmx duplicate-domains="true" domain="org.infinispan.plidm" mbean-server-lookup="org.infinispan.jmx.PlatformMBeanServerLookup"/>
-    <transport cluster="${jboss.partition.name:DefaultPartition}-idm" node-name="${jboss.node.name:gtn}" lock-timeout="240000"
-    <replicated-cache-configuration name="xml-configured-cache" mode="ASYNC">
-      <!-- JTA configuration. transactionManagerLookupClass will be configured programmatically to be:
-     - "org.infinispan.transaction.lookup.JBossTransactionManagerLookup" in JBoss AS environment
-     - "org.infinispan.transaction.lookup.JBossStandaloneJTAManagerLookup" in non-JBoss environment -->
-      <transaction auto-commit="true" locking="OPTIMISTIC" mode="BATCH"/>
-      <expiration interval="5000" lifespan="1800000"/>
-      <eviction strategy="LRU" size="100000"/>
-      <state-transfer timeout="60000" enabled="true"/>
-    </replicated-cache-configuration>
-
-  </cache-container>
-</infinispan>
