+++ web/portal/src/main/webapp/WEB-INF/conf/portal/controller-configuration.xml
@@ -163,17 +163,6 @@
       <set-method>register</set-method>
       <type>org.exoplatform.portal.application.StaticResourceRequestHandler</type>
     </component-plugin>
-    <component-plugin>
-      <name>StandaloneAppRequestHandler</name>
-      <set-method>register</set-method>
-      <type>org.exoplatform.portal.application.StandaloneAppRequestHandler</type>
-      <init-params>
-        <value-param>
-          <name>webui.configuration</name>
-          <value>app:/WEB-INF/standalone-webui-configuration.xml</value>
-        </value-param>
-      </init-params>
-    </component-plugin>
   </external-component-plugins>
 
 </configuration>
