--- web/portal/src/main/webapp/WEB-INF/jetty-web.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0"  encoding="ISO-8859-1"?>
-<!DOCTYPE Configure PUBLIC "-//Mort Bay Consulting//DTD Configure//EN" "http://jetty.mortbay.org/configure.dtd">
-<Configure class="org.mortbay.jetty.webapp.WebAppContext">
-
-  <Get id="serverObject" name="server"/>
-
-  <New id="jettySetup" class="org.gatein.wci.jetty.Jetty6Handler">
-    <Arg>
-      <Ref id="serverObject"/>
-    </Arg>
-  </New>
-
-</Configure>
-
-
