+++ web/portal/src/main/webapp/WEB-INF/jsp/externalRegistration/reset_password.jsp
