+++ web/portal/src/main/webapp/WEB-INF/jsp/onboarding/token_expired.jsp
