--- web/portal/src/main/webapp/login/jsp/keepalive.jsp
@@ -1,7 +1,0 @@
-<!--
-
-file:keepalive.jsp
-notice there's nothing here because the page doesn't
-do anything :-)
-
--->
