+++ web/portal/src/main/webapp/skin/externalRegistration.css
@@ -37,7 +37,7 @@ body {
 
 .alertSuccessMessage {
   padding: 10px;
-  background-color: #6CCDB0;;
+  background-color: #6CCDB0;
   color: #3C806A;
   border-radius: 8px;
   border: 2px solid #3C806A;
