--- web/portal/src/main/webapp/templates/groovy/webui/component/UIHomePagePortlet.gtmpl
@@ -1,99 +1,0 @@
-<%
-  import org.exoplatform.portal.webui.util.Util;
-  import org.exoplatform.portal.application.PortalRequestContext;
-
-	PortalRequestContext pcontext = Util.getPortalRequestContext();
-	String initialURI = pcontext.getInitialURI();
-%>
-
-<div class="UIHomePagePortlet ClearFix" id="$uicomponent.id">
-	<div class="TLContainer">
-		<div class="PortletDecoration">
-
-			<div class="HomePortletAdBackround">
-				<div class="AdImageLeft">
-                    <p class="slogan"><%=_ctx.appRes("UIHomePagePortlet.Label.BackgroundSlogan")%></p>
-					<div class="AdImageRight">
-					</div>
-				</div>
-			</div>
-
-			<div class="HomePortletContent">
-				<div class="LeftAccountsContainer">
-					<div class="RightAccountsContainer">
-						<div class="MiddleAccountsContainer">
-							<div class="InstructionTitle"><%=_ctx.appRes("UIHomePagePortlet.Label.Title")%></div>
-							<div class="AccountsContainerDeco ClearFix">
-								<div class="AccountBlock AdministratorUser">
-									<div class="AccountInfos">
-										<div class="AccountTitle"><a href="${_ctx.getPortalContextPath()}/login?username=root&amp;password=gtn&amp;initialURI=<%=initialURI%>" title="Access as root"><%=_ctx.appRes("UIHomePagePortlet.Label.Administrator")%></a></div>
-										<div class="Username ClearFix">
-											<div class="Lable"><%=_ctx.appRes("UIHomePagePortlet.Label.Username")%></div><span>root</span>
-										</div>
-										<div class="Passwords ClearFix">
-											<div class="Lable"><%=_ctx.appRes("UIHomePagePortlet.Label.Password")%></div><span>gtn</span>
-										</div>
-									</div>
-								</div>
-								<div class="SeparatorLine"><span></span></div>
-								<div class="AccountBlock ManagerUser">
-									<div class="AccountInfos">
-										<div class="AccountTitle"><a href="${_ctx.getPortalContextPath()}/login?username=john&amp;password=gtn&amp;initialURI=<%=initialURI%>" title="Access as john"><%=_ctx.appRes("UIHomePagePortlet.Label.Manager")%></a></div>
-										<div class="Username ClearFix">
-											<div class="Lable"><%=_ctx.appRes("UIHomePagePortlet.Label.Username")%></div><span>john</span>
-										</div>
-										<div class="Passwords ClearFix">
-											<div class="Lable"><%=_ctx.appRes("UIHomePagePortlet.Label.Password")%></div><span>gtn</span>
-										</div>
-									</div>
-								</div>
-								<div class="SeparatorLine"><span></span></div>
-								<div class="AccountBlock NormalUser">
-									<div class="AccountInfos">
-										<div class="AccountTitle"><a href="${_ctx.getPortalContextPath()}/login?username=mary&amp;password=gtn&amp;initialURI=<%=initialURI%>" title="Access as mary"><%=_ctx.appRes("UIHomePagePortlet.Label.User")%></a></div>
-										<div class="Username ClearFix">
-											<div class="Lable"><%=_ctx.appRes("UIHomePagePortlet.Label.Username")%></div><span>mary</span>
-										</div>
-										<div class="Passwords ClearFix">
-											<div class="Lable"><%=_ctx.appRes("UIHomePagePortlet.Label.Password")%></div><span>gtn</span>
-										</div>
-									</div>
-								</div>
-								<div class="SeparatorLine"><span></span></div>
-								<div class="AccountBlock DemoUser" style="margin-right: 0px;">
-									<div class="AccountInfos">
-										<div class="AccountTitle"><a href="${_ctx.getPortalContextPath()}/login?username=demo&amp;password=gtn&amp;initialURI=<%=initialURI%>" title="Access as demo"><%=_ctx.appRes("UIHomePagePortlet.Label.Demo")%></a></div>
-										<div class="Username ClearFix">
-											<div class="Lable"><%=_ctx.appRes("UIHomePagePortlet.Label.Username")%></div><span>demo</span>
-										</div>
-										<div class="Passwords ClearFix">
-											<div class="Lable"><%=_ctx.appRes("UIHomePagePortlet.Label.Password")%></div><span>gtn</span>
-										</div>
-									</div>
-								</div>
-                                                                <div class="ClearFix"></div>
-							</div>
-						</div>
-					</div>
-				</div>
-			</div>
-
-		</div>
-	</div>
-	<div class="TRContainer">
-		<div class="PortletDecoration">
-                	<div class="Information">
-				<div class="GuideText"><%=_ctx.appRes("UIHomePagePortlet.Label.IntroText")%></div>
-				<a class="VersionIcon" href="http://www.jboss.org/gatein/" target="_blank" title="Goto GateIn" ><span class="skipHidden">Goto GateIn</span></a>
-			</div>
-			<div class="DotLine"><span></span></div>
-                        <div class="ContactUs">
-				<div class="GuideText"><%=_ctx.appRes("UIHomePagePortlet.Label.GuideText")%></div>
-				<a class="ContactIcon" href="http://community.jboss.org/en/gatein?view=discussions" target="_blank" title="Goto Discussions"><span class="skipHidden">Goto Discussions</span></a>
-			</div>
-		</div>
-		<div class="ClearFix"></div>
-	</div>
-	<div class="ClearFix"></div>
-</div>
-
