--- web/portal/src/main/webapp/templates/skin/webui/component/UIHomePagePortlet/DefaultStylesheet.css
@@ -1,263 +1,0 @@
-/**
- * Copyright (C) 2009 eXo Platform SAS.
- * 
- * This is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- * 
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- * 
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
- */
-
-.UIWindow .UIHomePagePortlet {	
-	margin: 0px;
-	padding: 10px;
-}
-
-.UIHomePagePortlet {	
-	background: white;
-	padding: 7px 8px 7px 7px;
-}
-
-.UIHomePagePortlet .TRContainer .PortletDecoration {	
-	padding: 10px;
-}
-
-.UIHomePagePortlet .TRContainer {
-	width: 225px;
-	height: 385px;
-	float: right; /* orientation=lt */
-	float: left; /* orientation=rt */
-	border: 1px solid #eee;	
-	text-align: center;
-	background: url('DefaultSkin/background/BgRightColumn.gif') repeat-x bottom center;
-	margin-top:1px;
-}
-
-.UIHomePagePortlet .TRContainer a {
-	text-decoration: none;
-}
-
-.UIHomePagePortlet .TRContainer .GuideText {
-	text-align: justify;
-	color: #777777;
-}
-
-.UIHomePagePortlet .TRContainer .DotLine {
-	background: url("DefaultSkin/background/Line.gif") no-repeat -2px top;
-	height: 1px;
-	width: 182px;
-	margin: 7px auto;
-}
-
-.UIHomePagePortlet .TRContainer .VersionIcon {
-	background: url("DefaultSkin/background/HomePagePortlet.gif") no-repeat left -978px;
-	height: 65px;
-	width: 80px;
-	margin: 20px auto;
-	cursor: pointer;
-	display: block;
-}
-
-.UIHomePagePortlet .TRContainer .ContactIcon {
-	background: url("DefaultSkin/background/HomePagePortlet.gif") no-repeat left -1209px;
-	height: 70px;
-	width: 80px;
-	margin: 20px auto;
-	cursor: pointer;
-	display: block;
-}
-
-.UIHomePagePortlet  .TLContainer {
-	border: 1px solid #ffffff;
-	float: left; /* orientation=lt */	
-	float: right; /* orientation=rt */
-	width: 713px;
-}
-
-.UIHomePagePortlet .TLContainer .PortletDecoration {	
-	background: #fff;
-}
-
-.UIHomePagePortlet .HomePortletAdBackround {	
-	background: #fff url('DefaultSkin/background/BgRepeat.gif') repeat-x top;
-}
-
-.UIHomePagePortlet .AdImageLeft {
-        background: url('DefaultSkin/background/background-rtl.jpg') no-repeat right top; /* orientation=rt */
-        background: url('DefaultSkin/background/background-ltr.jpg') no-repeat left top; /* orientation=lt */
-}
-
-.UIHomePagePortlet .AdImageLeft {
-        position: relative;
-}
-
-.UIHomePagePortlet .AdImageRight {
-	text-align: right; /* orientation=lt */
-	text-align: left; /* orientation=rt */
-	height: 255px; 
-	padding: 0px 10px 0px 200px; /* orientation=lt */
-	padding: 0px 200px 0px 10px; /* orientation=rt */
-}
-
-.UIHomePagePortlet .slogan {
-	position: absolute;
-	font-size: 25px;
-	line-height: 29px;
-	font-weight: 100;
-	right: 50px; /* orientation=rt */
-	left: 58px; /* orientation=lt */
-	top: 134px;
-	letter-spacing: -0.4px;
-	width: 280px;
-	font-family: "Frutiger LT Std", Frutiger, Helvetica, Arial, Sans-serif;
-	color: #7b7b7b;
-        /*Disable for now since it does not gracefully degrade and does not have enough browser support (especially on mobile browsers */
-	/*background: -webkit-linear-gradient(#b9b9b9, #535353); */ /* gradient effect only works on webkit */
-	/*-webkit-background-clip: text;
-	-webkit-text-fill-color: transparent; */
-}
-
-.UIHomePagePortlet .AdImageLeft .EmptyBlock {
-	height: 120px;
-	font-size: 21px;
-	color: #616161;
-	text-align: center;
-	font-family: arial;
-	font-weight: bold;
-	float: right; /* orientation=lt */
-	float: left; /* orientation=rt */
-	padding: 50px 0 0;
-}
-
-.UIHomePagePortlet .AdImageLeft .EmptyBlock div {
-	font-size: 19px;
-}
-
-.UIHomePagePortlet .AdImageLeft .ProductSlogan {
-	font-family: arial;
-	color: #666666;
-	float: right;
-	font-size: 9px;
-	width: 500px;
-}
-
-.UIHomePagePortlet .AdImageLeft .ProductIntro {
-	font-size: 14px;
-	color: #315d8f;
-	line-height: 16px;
-}
-
-.UIHomePagePortlet .HomePortletContent {
-	margin-top: 5px;
-}
-
-.UIHomePagePortlet .HomePortletContent .LeftAccountsContainer {
-}
-
-_html .UIHomePagePortlet .HomePortletContent .LeftAccountsContainer {
-	background: url('DefaultSkin/background/HomePagePortlet.gif') no-repeat 4px -529px;
-}
-
-.UIHomePagePortlet .HomePortletContent .RightAccountsContainer {
-}
-
-_html .UIHomePagePortlet .HomePortletContent .RightAccountsContainer {
-	background: url('DefaultSkin/background/HomePagePortlet.gif') no-repeat 114% -799px;/* orientation=rt */
-	padding-right: 8px;/* orientation=rt */
-}
-
-.UIHomePagePortlet .HomePortletContent .MiddleAccountsContainer {
-	background: url('DefaultSkin/background/HomePagePortlet.gif') repeat-x center -379px;
-	height: 127px;
-	padding: 0px 8px;
-}
-
-.UIHomePagePortlet .HomePortletContent .AccountsContainerDeco {
-	height: 105px;
-	overflow: hidden;
-}
-
-.UIHomePagePortlet .HomePortletContent .InstructionTitle {
-	font-size: 12px;
-	line-height: 22px;
-	color: #808185;
-	text-align: left; /* orientation=lt */
-	text-align: right; /* orientation=rt */
-	position:relative;
-	top: 4px;
-}
-
-.UIHomePagePortlet .HomePortletContent .AccountBlock {
-	height: 105px; 
-	width: 160px;
-	float: left; /* orientation=lt */
-	float: right; /* orientation=rt */
-	text-align: right; /* orientation=rt */
-}
-
-.UIHomePagePortlet .HomePortletContent .AdministratorUser {
-	background: url('DefaultSkin/background/HomePagePortlet.gif') no-repeat right top; /* orientation=lt */
-	background: url('DefaultSkin/background/HomePagePortlet-rt.gif') no-repeat 10px top; /* orientation=rt */
-}
-
-.UIHomePagePortlet .HomePortletContent .ManagerUser {
-	background: url('DefaultSkin/background/HomePagePortlet.gif') no-repeat right -95px; /* orientation=lt */
-	background: url('DefaultSkin/background/HomePagePortlet-rt.gif') no-repeat 10px -95px; /* orientation=rt */
-}
-
-.UIHomePagePortlet .HomePortletContent .NormalUser {
-	background: url('DefaultSkin/background/HomePagePortlet.gif') no-repeat right -185px; /* orientation=lt */
-	background: url('DefaultSkin/background/HomePagePortlet-rt.gif') no-repeat 10px -185px; /* orientation=rt */
-}
-
-.UIHomePagePortlet .HomePortletContent .DemoUser {
-	background: url('DefaultSkin/background/HomePagePortlet.gif') no-repeat right -275px; /* orientation=lt */
-	background: url('DefaultSkin/background/HomePagePortlet-rt.gif') no-repeat 10px -275px; /* orientation=rt */
-}
-
-.UIHomePagePortlet .HomePortletContent .SeparatorLine {
-	height: 100px;
-	width: 9px;
-	float: left; /* orientation=lt */
-	float: right; /* orientation=rt */
-	padding: 0px 4px 0px;
-	background: url('DefaultSkin/background/HomePagePortlet.gif') no-repeat left -675px; /* orientation=lt */
-	background: url('DefaultSkin/background/HomePagePortlet-rt.gif') no-repeat right -675px; /* orientation=rt */
-}
-
-.UIHomePagePortlet .HomePortletContent .AccountInfos {
-	padding-top: 8px;
-}
-
-.UIHomePagePortlet .AccountInfos .Lable {
-	line-height: 16px;
-	float: left; /* orientation=lt */
-	float: right; /* orientation=rt */
-	display: block;
-}
-
-.UIHomePagePortlet .AccountInfos span {
-	font-weight: bold;
-	color: #2c6fbd;
-	padding: 0px 3px;
-	line-height: 16px;
-	float: left; /* orientation=lt */
-	float: right; /* orientation=rt */
-	display: block;
-}
-
-.UIHomePagePortlet .HomePortletContent .AccountTitle {
-	font-size: 12px;
-	color: #454e56;
-	line-height: 40px;
-	font-weight: bold;
-}
