+++ webui/portal/src/main/java/org/exoplatform/portal/webui/application/UIStandaloneAppContainer.java
@@ -23,7 +23,7 @@ import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 
 import org.exoplatform.portal.application.StandaloneAppRequestContext;
-import org.exoplatform.web.login.LoginServlet;
+import org.exoplatform.web.login.LoginUtils;
 import org.exoplatform.web.login.LogoutControl;
 import org.exoplatform.web.security.security.AbstractTokenService;
 import org.exoplatform.web.security.security.CookieTokenService;
@@ -53,19 +53,19 @@ public class UIStandaloneAppContainer extends UIContainer {
                 AbstractTokenService tokenService = AbstractTokenService.getInstance(CookieTokenService.class);
                 tokenService.deleteToken(token);
             }
-            token = LoginServlet.getOauthRememberMeTokenCookie(req);
+            token = LoginUtils.getOauthRememberMeTokenCookie(req);
             if (token != null) {
                 AbstractTokenService tokenService = AbstractTokenService.getInstance(CookieTokenService.class);
                 tokenService.deleteToken(token);
             }
 
             LogoutControl.wantLogout();
-            Cookie cookie = new Cookie(LoginServlet.COOKIE_NAME, "");
+            Cookie cookie = new Cookie(LoginUtils.COOKIE_NAME, "");
             cookie.setPath(req.getContextPath());
             cookie.setMaxAge(0);
             context.getResponse().addCookie(cookie);
 
-            Cookie oauthCookie = new Cookie(LoginServlet.OAUTH_COOKIE_NAME, "");
+            Cookie oauthCookie = new Cookie(LoginUtils.OAUTH_COOKIE_NAME, "");
             oauthCookie.setPath(req.getContextPath());
             oauthCookie.setMaxAge(0);
             context.getResponse().addCookie(oauthCookie);
@@ -77,7 +77,7 @@ public class UIStandaloneAppContainer extends UIContainer {
             Cookie[] cookies = req.getCookies();
             if (cookies != null) {
                 for (Cookie cookie : cookies) {
-                    if (LoginServlet.COOKIE_NAME.equals(cookie.getName())) {
+                    if (LoginUtils.COOKIE_NAME.equals(cookie.getName())) {
                         return cookie.getValue();
                     }
                 }
